+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,7 +1,17 @@
-# v.0.19.0 (2021-09-12)
+# v.0.20.0 (2021-09-12)
 
 - upgraded "cosmos-client": "0.42.14"
 - changed class Client to extend BaseXChainClient
+
+# v.0.19.0 (2021-09-10)
+
+## Breaking change
+
+- Extract `buildDepositTx` from `Client` into `utils`
+
+## Add
+
+- Add `getBalance` to `utils`
 
 # v.0.18.0 (2021-09-08)